Transfer Config Merge - avoid Artifactory min version check (#785)
